2nd Get in WriterController accepts Long

+++ Controllers/WriterController.cs
@@ -31,7 +31,7 @@ namespace IndyBooks.Controllers
 
         // GET: api/Writer/5
         [HttpGet("{id}", Name = "Get")]
-        public IActionResult Get(int id)
+        public IActionResult Get(long id)
         {
             return "value";
         }
